requirements.txt
pyautogui 
opencv-python  
pyaudio 
moviepy
keyboard 
numpy
pillow